Adding config prop to CrossmintPayButton
