--- packages/ui/react-ui/src/hooks/types.ts
@@ -1,13 +1,0 @@
-export enum clientNames {
-    reactUi = "client-sdk-react-ui",
-}
-
-export enum baseUrls {
-    prod = "https://www.crossmint.io",
-    dev = "http://localhost:3001",
-}
-
-export enum customHeaders {
-    clientVersion = "X-Client-Version",
-    clientName = "X-Client-Name",
-}
