+++ packages/ui/react-ui/src/hooks/useCrossMintModal.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { LIB_VERSION } from "../version";
-import { baseUrls, clientNames } from "./types";
+import { baseUrls, clientNames } from "../types";
 
 function createPopupString() {
     return `height=750,width=400,left=${window.innerWidth / 2 - 200},top=${
